Resolve PIVOT-569 for ListView and TableView.
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@986435 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk-terra/src/org/apache/pivot/wtk/skin/terra/TerraCalendarSkin.java
@@ -521,7 +521,7 @@ public class TerraCalendarSkin extends CalendarSkin {
             throw new RuntimeException(exception);
         }
 
-        weekdayCharacterIndex = JSON.get(resources, "weekdayCharacterIndex");
+        weekdayCharacterIndex = (Integer)JSON.get(resources, "weekdayCharacterIndex");
     }
 
     @Override
